Move clar module into resonance
Since Clar structure generation will be a part of resonance
  structure generation
Also merge clarTest into resonanceTest
